Bump the compiler service over 35 and almost to 38.
35 is where a lot of APIs changed. The main culprits seem to be
https://github.com/dotnet/fsharp/commit/78bb83aca6cba3d85dee111211d4c0c99a37595d
and https://github.com/dotnet/fsharp/pull/8805.

This commit fails in many of the same ways, but 38 goes with F# and .NET 5,
which are not quite targetable by the project upgrader right now, so I'll go in
stages.

The hope is that according to https://github.com/dotnet/fsharp/issues/6161 and
https://github.com/dotnet/fsharp/issues/6535 the mscorlib errors will
eventually be solved.