Get Kl.Make compiling.
Upgraded to 6.0, was guided by missing assembly errors in `dotnet run --project
Kl.Make`.

There are warnings about the mismatch between FCS 38 and F# 6, but

a) The projects are tracked together so it's arguably better to pin the latter.
b) There are a bunch of changes between 38 and the oldest 6-friendly version,
41.