Target net6.0 with the Shen* projects.
The compilation process appears to be broken indeed. There are a whole bunch of
ununsed warnings during compile and no Shen code runs.

https://github.com/dotnet/fsharp/blob/main/release-notes.md#fsharp-compiler-service-3900
is the next FCS upgrade to
handle. https://fsharp.github.io/fsharp-compiler-docs/fsharp-core-notes.html
suggests that the compiler/runtime mismatch is less than ideal but shouldn't be
a showstopper.