Target 38.1.
As written, the version from GitHub doesn't work and is a little different than
the zip package on the Shen website. The latter is what works better: It
doesn't need dict and doesn't contain initialise any more. 38.2 should work
about the same. Since the paths have not changed, you will need to extract the
files to kernel and rename their KLambda directory to klamnbda.