Ensure non-alphanumeric interns are not symbols.
The internal `Sym` structure can hold arbitrary strings just fine, but the
behavior of `symbol?` broke this test.

While the string "fsdf{}.$%2" can be `intern`ed, the Shen source code for
`symbol?` specifies that symbols contain only alphanumerics. Implementations
differ in how they handle this. Shen/Scheme 39 overrides `symbol?` and
`analyse-symbol?` such that `(intern "fsdf{}.$%2")` is a symbol. The SBCL Shen
distribution 39.1 as compiled under SBCL 2.4.5 on linux-x86_64 can `intern` the
string but does not recognize it as a symbol. We choose to follow this behavior
with the potential of creating an override later.